pyproject.toml
[tool.poetry]
name = "meme_khani_api"
version = "0.1.0"
description = ""
authors = ["hemanta212 <sharmahemanta.212@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
flask = "^1.1.2"
flask-sqlalchemy = "^2.4.4"
PyJWT = "^1.7.1"
python-dotenv = "^0.14.0"
mysql-connector-python = "^8.0.21"

[tool.poetry.dev-dependencies]
pylint = "^2.6.0"
python-language-server = "^0.35.1"
black = "^20.8b1"
ptpython = "^3.0.7"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
